backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
cryptography==44.0.0
httpx==0.28.1
python-multipart==0.0.20
openpyxl==3.1.5
xlrd==2.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
sentry-sdk[fastapi]==2.19.2
